Fix: Add ViewModel folder and move AuthView.ViewModel to an extension in that folder
- Free scoping
